+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.7.1 [unreleased]
 
 ### Features
 -   [#277](https://github.com/latera/homs/pull/277) Allow to configure widget polling interval.
+-   [#275](https://github.com/latera/homs/pull/275) Set logging level from env file.
 
 v1.7.0 [2019-04-25]
 -------------------